Refactor (#64)
* Adding in enable access

* Adding in file support

* Ran rubocop

* Adding in tests

* Fixing small errors

* Responding to comments

* Fixing error

* Refactors the main action's instance variables

* Moves paths and upload_token formatter to helper

* Small change to get CI passing

* Moves methods from helper to api client. Spec file is split up into helper and api client spec files

* Change api_client from module to class

* Converts api_client class methods to instance methods. Makes internal methods private

* Fixed connection url

* Refactors fake binary path/contents
