+++ lib/fastlane/plugin/firebase_app_distribution/helper/firebase_app_distribution_helper.rb
@@ -1,16 +1,13 @@
 require 'fastlane_core/ui/ui'
-require 'cfpropertylist'
-
 module Fastlane
   UI = FastlaneCore::UI unless Fastlane.const_defined?("UI")
-
   module Helper
     module FirebaseAppDistributionHelper
-      def get_value_from_value_or_file(value, path)
+      def self.get_value_from_value_or_file(value, path)
         if (value.nil? || value.empty?) && !path.nil?
           begin
             return File.open(path).read
-          rescue
+          rescue Errno::ENOENT
             UI.crash!("#{ErrorMessage::INVALID_PATH}: #{path}")
           end
         end
@@ -20,13 +17,13 @@ module Fastlane
       # Returns the array representation of a string with comma seperated values.
       #
       # Does not work with strings whose individual values have spaces. EX "Hello World" the space will be removed to "HelloWorld"
-      def string_to_array(string)
+      def self.string_to_array(string)
         return nil if string.nil? || string.empty?
         string_array = string.gsub(/\s+/, '').split(",")
         return string_array
       end
 
-      def get_ios_app_id_from_archive(path)
+      def self.get_ios_app_id_from_archive(path)
         app_path = parse_plist("#{path}/Info.plist")["ApplicationProperties"]["ApplicationPath"]
         UI.shell_error!("can't extract application path from Info.plist at #{path}") if app_path.empty?
         identifier = parse_plist("#{path}/Products/#{app_path}/GoogleService-Info.plist")["GOOGLE_APP_ID"]
